# SIMD: SSE2 intrinsic implementation for float64 input of np.enisum