@Qiyu8 2020-06-10
> Not sure why aarch64 failed the first time around and passed the second time. Hmm...

That's strange, because the code that changes here only affects the X86 platform.