@charris 2020-06-10
> changes here only affects the X86 platform

I'm suspecting OpenBLAS, but it could be something else.